Less warnings (#813)

+++ libp2p/muxers/yamux/yamux.nim
@@ -84,7 +84,7 @@ proc `$`(header: YamuxHeader): string =
 proc encode(header: YamuxHeader): array[12, byte] =
   result[0] = header.version
   result[1] = uint8(header.msgType)
-  result[2..3] = toBytesBE(cast[uint16](header.flags))
+  result[2..3] = toBytesBE(uint16(cast[uint8](header.flags))) # workaround https://github.com/nim-lang/Nim/issues/21789
   result[4..7] = toBytesBE(header.streamId)
   result[8..11] = toBytesBE(header.length)
